security: introduce new option DECOUPLE_ROOTFS_FROM_RAMDISK
By default, the kernel ramdisk only mount rootfs that are built
together. Introduce DECOUPLE_ROOTFS_FROM_RAMDISK to decouple that

Resolves: QTI9X07-4650
Related: QTI9X07-3302

+++ meta-swi-mdm9x28/recipes-core/initrdsripts/initramfs-mdminit.bbappend
@@ -1,10 +1,12 @@
 
 do_install() {
     if [[ "${DM_VERITY_ENCRYPT}" = "on" ]]; then
-        local dm_hash_filename="${DEPLOY_DIR_IMAGE}/rootfs.rhash"
-        #skip signature
-        ROOTHASH=$(dd if=$dm_hash_filename bs=1 count=64 2>/dev/null)
-        sed -i 's/^.*ROOTHASH=.*$/ROOTHASH='${ROOTHASH}'/g' ${WORKDIR}/init.sh
+        if [[ "${DECOUPLE_ROOTFS_FROM_RAMDISK}" = "off" ]]; then
+             local dm_hash_filename="${DEPLOY_DIR_IMAGE}/rootfs.rhash"
+             #skip signature
+             ROOTHASH=$(dd if=$dm_hash_filename bs=1 count=64 2>/dev/null)
+             sed -i 's/^.*ROOTHASH=.*$/ROOTHASH='${ROOTHASH}'/g' ${WORKDIR}/init.sh
+        fi
     fi
 
     install -m 0755 ${WORKDIR}/init.sh ${D}/init
